Update jquery.msgBox.js
In the functin show() there is a statement $(window).bind("resize", function (e) { ... }) to register a resize event handler to the window but no corresponding statement $(window).unbind("resize")  in the function hide(). If we create new msgBox and close it and then repeat this operation for several times the DOM count will increase gradually.